requirements.txt
numpy
pillow
torch
torchvision
scikit-image
gradio